Adds beginnings of movement system
Adds on click function for pawns to show valid potential
movement locations.

This movement function checks to make sure the move space is one
contained within the bounds of the board and if is valid with color the
space yellow.

For a pawn specifically the function can tell if it the given pawns
first move thus allowing for an additional space of movement. It also
checks the spaces diagonally forward of it to detect if an enemy piece
is present to allow for a legal diagonal movement.

Pieces cannot currently take other pieces.

If once a pawn is selected and a valid space is selected after that the
piece will migrate to the new location inheriting its information and
clearing the information of its previous location, leaving it empty.

README file is NOT updated yet.