Adds King movement, reworks validation
Adds movement to kings using nested for...loops.

Reworks validLocation function entirely to a system that uses a
temporary array to push valid space ids to and them finally assigning
them back to the given array. This method is able to more thoroughly
check the array for invalid locations.

Updates the README file to reflect current status of the project.