Adds check system and begins checkMate system
Adds check system that uses arrays to store all possible moves for
either side to check for the enemy king.

Adds beginnings of checkMate system that builds off of the check system
but this is not functional yet.

Reworks several variables and adds "if" checks to certain functions to
allow them to work with the check/checkMate systems.

Updates README file to reflect current status of the project.