I'm back baby! Check system complete overhaul
Fixes bug where en passant caused path blocking for certain pieces

Adds correct functionality for checkForCheck system to correctly
determine when a players king is in check.

Fixes a checked kings ability to correctly determine a path of escape
out of a check if one is possible.

Completely overhauls checkMate system.

Adds ability to detect if friendly pieces can block for a king in check.

Disables game once victory is achieved.

Adds end game message, dependent on the victor. **This has additions to
both the index and the stylesheet as well.

Adds beginning of a turn based Draw system. Added so far are a turn
countdown and a message once it is reached.